my_agent/requirements.txt
langchain_huggingface
deeplake
langgraph-supervisor
langchain-groq
langgraph
langchain-community
sentence-transformers
langchain-google-genai